+++ app/src/components/__tests__/TeamDetails.spec.ts
@@ -1,7 +1,7 @@
 // TeamDetails.spec.ts
 import { it, expect, describe, vi, beforeEach } from 'vitest'
 import { mount } from '@vue/test-utils'
-import TeamDetails from '../TeamDetails.vue'
+import TeamDetails from '../sections/SingleTeamView/Team/TeamDetails.vue'
 
 describe('TeamDetails.vue', () => {
   let wrapper: ReturnType<typeof mount>
