+++ app/src/components/sections/SingleTeamView/Governance/ProposalDetails.vue
