+++ app/src/components/sections/SingleTeamView/Team/MemberCard.vue
