+++ app/src/components/sections/SingleTeamView/Team/TeamDetails.vue
