+++ app/src/components/sections/TeamView/TeamCard.vue
